More RustDoc tweaks for the sign module, and restore the crypto common module by moving keys/keypair.rs to crypto/common.rs.
