+++ src/sign/crypto/common.rs
