+++ src/sign/crypto/mod.rs
@@ -1,2 +1,3 @@
+pub mod common;
 pub mod openssl;
 pub mod ring;
