+++ src/sign/keys/mod.rs
@@ -1,5 +1,4 @@
 pub mod bytes;
 pub mod keymeta;
-pub mod keypair;
 pub mod keyset;
 pub mod signingkey;
